Add files via upload
The user is asked for midterm and final exam grades. The average is calculated with 40% of the midterm grade and 60% of the final grade and the result is shown on the screen. The user is asked if he/she wants to know the letter grade. If the user click the "OK" button, the letter grade and whether the user has passed or failed the course is displayed on the screen. If the user click the "Cancel" button, the process ends and the letter grade is not displayed.